heavy optimizations - inline often called Map class functions that are pretty small.
Change the way we generate the game tree, instead of allocating nodes individually, use a deque to act as a simple memory pool for efficient allocation and deallocation.

Instead of saving a new Map for each node in the game tree, generate the current map on the fly as we traverse through it.